Use full localModList when determining mod's dependencies
Previously modifiableModList was used. It was derived from localModList
by sorting it. However, one of the sort options allows hiding disabled
mods, meaning that the list might be missing an arbitrary number of
mods.

Instead use localModList which is guaranteed to include all the
dependencies/dependants. This will change the functionality a bit by
possibly changing the order of mods returned by methods that return
lists of dependencies and dependants. Given how these results are used,
I *assume* it won't matter that much for the rest of the functionality
(since the order could already be quite arbitrary) or the user.