Add VuexProvider
This allows us to access the current instance of Vuex store outside of
Vue components, both in R2MM and TSMM.

Since this is a very simple case, defining the abstract interface and
implementation separately were skipped. The provider pattern was
mimicked for consistency's sake. I'm not sure if this is the best
approach, but it should work now and will be rather easy to change
later.